Refactor the app code without global variables
No functionality is changed, we simply wrap the code that modifies the
app (e.g. adding routes) into funtions.
This way the app is not built at import time but explicitely in the
entry point. This will make for cleaner dependency injection, in
particular for the upcoming objets required for authentification.